Update navigation3 to latest snapshot
